Add new methods to support 'launch-until' workflow
Two methods are added to the WorflowRunner interface:
(1) cancelTaskTree(), which cancels a task and all of its children
(2) isTaskDone(), is similar to isTaskComplete() but also allows the
workflow to address task error conditions.

Both of these features are added to support a new dynamic
launch-task-until workflow, which will appear in a new demo.