Reduce infrequent interpreter shutdown errors
The workflow could previously panic during normal shutdown due to the
TaskManager thread attempting to close at the same time as the python
interpreter. This led to a confusing error message even though the
run was successful. Rate of occurance seems to have been
approx 1 in 30 runs. With this change the workflow waits for the
TaskRunner and the TaskManager threads to terminate (under normal
conditions) before allowing the primary workflow thread to terminate.