Remove a race condition in task processing
When determining if a task has completed, the threading event
setup to indicate task completion and the tasks's thread status
were checked in the wrong order, allowing for an erroneous failed
task report. The occurance rate for this error was believed to be
very infrequent. This patch should eliminate the race condition.